Adds time layout to catmaply object

This layout is used if the x axis is of type time.

Usage

catmaply_time_layout(
  fig,
  df,
  x,
  x_order,
  x_side,
  x_tickangle,
  x_range,
  y,
  y_order,
  y_side,
  y_tickangle,
  tickformatstops,
  font_family,
  font_size,
  font_color,
  legend,
  rangeslider
)

Arguments

fig

plotly object

df

data.frame or tibble holding the data.

x

column name holding the axis values for x.

x_order

column name holding the ordering axis values for x. if no order is specified, then x will be used for ordering x; (default:"x").

x_side

on which side the axis labels on the x axis should appear. options: c("top", "bottom"); (default:"top").

x_tickangle

the angle of the axis label on the x axis. options: range -180 until 180; (default:90).

x_range

the initial range that should be displayed on the x axis. Only works with non-time x-axis at the moment; (default: 30).

y

column name holding the axis values for y.

y_order

column name holding the ordering axis values for y. if no order is specified, then y will be used for ordering y; (default:"y").

y_side

on which side the axis labels on the y axis should appear. options: c("left", "right"); (default:"left").

y_tickangle

the angle of the axis label on the x axis. options: range -180 until 180; (default:0).

tickformatstops

used only if x axis is of type c("POSIXct", "POSIXt"). List of named list where each named list has one or more of the keys listed here: https://plotly.com/r/reference/#heatmap-colorbar-tickformatstops. Default is optimized for summarized data of level day 24 hours; (default: list( list(dtickrange = list(NULL, 1000), value = " list(dtickrange = list(1000, 60000), value = " list(dtickrange = list(60000, 3600000), value = " list(dtickrange = list(3600000, 86400000), value = " list(dtickrange = list(86400000, 604800000), value = " list(dtickrange = list(604800000, "M1"), value = " list(dtickrange = list("M1", "M12"), value = " list(dtickrange = list("M12", NULL), value = " ) )

font_family

the typeface that will be applied by the web browser. The web browser will only be able to apply a font if it is available on the system which it operates. Provide multiple font families, separated by commas, to indicate the preference in which to apply fonts if they aren't available on the system; (default: c("Open Sans", "verdana", "arial", "sans-serif")).

font_size

font size to be used for plot. needs to be a number greater than or equal to 1; (default: 12).

font_color

font color to be used for plot; (default: "#444")

legend

boolean indicating if legend should be displayed or not; (default: TRUE).

rangeslider

whether or not the rangeslider should be displayed or not; (default: TRUE).

Value

plot_ly object